Refactor auth routes into separate files
