+++ src/api/auth/src/routes/index.js
@@ -1,0 +1,13 @@
+const { Router } = require('@senecacdot/satellite');
+const { errors } = require('celebrate');
+
+const router = Router();
+
+router.use('/login', require('./login'));
+router.use('/logout', require('./logout'));
+router.use('/sp', require('./saml-metadata'));
+
+// Let Celebrate handle validation errors
+router.use(errors());
+
+module.exports = router;
